Implement props and temp workaround building subfile resource key being wrong
Unsure what the issue is exactly right now, but it seems that reading an additional byte prior to the group id in Building causes the persistent resource key to be incorrect as that first byte happens to be part of that key.

Removing it causes an error to be thrown as the expected size is one byte smaller than the amount read. It seems this additional byte does exist somewhere, but not sure where at the moment. Will need to dig deeper on this...